BetterForwards: Another errorboundary
